updated delete functionality

+++ routers/projectsRouter/index.js
@@ -17,16 +17,23 @@ router.post("/", (req, res) => {
 // GET A PROJECT BY ID
 router.get("/:id", (req, res) => {
   const { id } = req.params;
+
+  // get project from database using data/helpers
   DB.getProject(id)
     .then(result => {
       result = result[0];
-      DB.getActions(id).then(actions => {
-        const project = Object.assign({}, result, { actions: actions });
-        res.json({ project });
-      });
+      // check if falsy, return error for bad ID
+      if (!result) {
+        res.status(404).json({ error: "This project does not exist" });
+      } else {
+        DB.getActions(id).then(actions => {
+          const project = Object.assign({}, result, { actions: actions });
+          res.json({ project });
+        });
+      }
     })
     .catch(err => {
-      res.status(500).send(err);
+      res.status(500).json({ error: "error code: NOGTPROJID" });
     });
 });
 
@@ -36,19 +43,32 @@ router.delete("/:id", (req, res) => {
   const { name } = req.body;
   const { id } = req.params;
   DB.getProject(id).then(project => {
-    DB.deleteProject(id, name)
-      .then(deletedProject => {
-        DB.deleteActionsForDeletedProject(id)
-          .then(deletedActions => {
-            res.json({ deletedActions, deletedProject });
+    DB.getActions(id).then(actions => {
+      // check if user matched names to be deleted
+      if (name === project[0].name) {
+        // search for and delete project itself
+        DB.deleteProject(id, name)
+          .then(deletedProject => {
+            // then search for and delete all actions related to project
+            DB.deleteActionsForDeletedProject(id)
+              .then(deletedActions => {
+                const finalProject = Object.assign({}, project, { actions });
+                // return info that shows what was deleted
+                res.json({ finalProject, deletedActions, deletedProject });
+              })
+              .catch(err => {
+                res.status(500).json({ error: "error code: NODELIDACT" });
+              });
           })
           .catch(err => {
-            res.status(500).json({ error: "error code: NODELIDACT" });
+            res.status(500).json({ error: "error code: NODELIDPROJ" });
           });
-      })
-      .catch(err => {
-        res.status(500).json({ error: "error code: NODELIDPROJ" });
-      });
+      } else {
+        res.status(401).json({
+          error: "Please enter the correct name of the project to delete"
+        });
+      }
+    });
   });
 });
